rename preview package to extractors
