+++ k9mail/src/main/java/com/fsck/k9/mailstore/LocalFolder.java
@@ -53,9 +53,9 @@ import com.fsck.k9.mail.internet.SizeAware;
 import com.fsck.k9.mail.message.MessageHeaderParser;
 import com.fsck.k9.mailstore.LockableDatabase.DbCallback;
 import com.fsck.k9.mailstore.LockableDatabase.WrappedException;
-import com.fsck.k9.message.preview.MessagePreviewCreator;
-import com.fsck.k9.message.preview.PreviewResult;
-import com.fsck.k9.message.preview.PreviewResult.PreviewType;
+import com.fsck.k9.message.extractors.MessagePreviewCreator;
+import com.fsck.k9.message.extractors.PreviewResult;
+import com.fsck.k9.message.extractors.PreviewResult.PreviewType;
 import com.fsck.k9.preferences.Storage;
 import com.fsck.k9.preferences.StorageEditor;
 import org.apache.commons.io.IOUtils;
