+++ k9mail/src/main/java/com/fsck/k9/message/extractors/TextPartFinder.java
@@ -1,4 +1,4 @@
-package com.fsck.k9.message.preview;
+package com.fsck.k9.message.extractors;
 
 
 import android.support.annotation.NonNull;
