+++ k9mail/src/test/java/com/fsck/k9/message/extractors/EncryptionDetectorTest.java
@@ -1,4 +1,4 @@
-package com.fsck.k9.message.preview;
+package com.fsck.k9.message.extractors;
 
 
 import com.fsck.k9.mail.Message;
