+++ k9mail/src/test/java/com/fsck/k9/message/extractors/MessagePreviewCreatorTest.java
@@ -1,10 +1,10 @@
-package com.fsck.k9.message.preview;
+package com.fsck.k9.message.extractors;
 
 
 import com.fsck.k9.mail.Message;
 import com.fsck.k9.mail.Part;
 import com.fsck.k9.mail.internet.MimeMessage;
-import com.fsck.k9.message.preview.PreviewResult.PreviewType;
+import com.fsck.k9.message.extractors.PreviewResult.PreviewType;
 import org.junit.Before;
 import org.junit.Test;
